requirements.txt
fastapi
uvicorn
python-telegram-bot
apscheduler
databases[postgresql]
psycopg2-binary
aiosqlite